+++ pymatgen/analysis/chemenv/coordination_environments/coordination_geometry_finder.py
@@ -180,6 +180,12 @@ class AbstractGeometry:
     @classmethod
     def from_cg(cls, cg, centering_type='standard',
                 include_central_site_in_centroid=False):
+        """
+        :param cg:
+        :param centering_type:
+        :param include_central_site_in_centroid:
+        :return:
+        """
         central_site = cg.get_central_site()
         bare_coords = [np.array(pt, np.float) for pt in cg.points]
         return cls(central_site=central_site, bare_coords=bare_coords,
@@ -187,41 +193,71 @@ class AbstractGeometry:
                    include_central_site_in_centroid=include_central_site_in_centroid)
 
     def points_wcs_csc(self, permutation=None):
+        """
+        :param permutation:
+        :return:
+        """
         if permutation is None:
             return self._points_wcs_csc
         return np.concatenate((self._points_wcs_csc[0:1], self._points_wocs_csc.take(permutation, axis=0)))
 
     def points_wocs_csc(self, permutation=None):
+        """
+        :param permutation:
+        :return:
+        """
         if permutation is None:
             return self._points_wocs_csc
         return self._points_wocs_csc.take(permutation, axis=0)
 
     def points_wcs_ctwcc(self, permutation=None):
+        """
+        :param permutation:
+        :return:
+        """
         if permutation is None:
             return self._points_wcs_ctwcc
         return np.concatenate((self._points_wcs_ctwcc[0:1], self._points_wocs_ctwcc.take(permutation, axis=0)))
 
     def points_wocs_ctwcc(self, permutation=None):
+        """
+        :param permutation:
+        :return:
+        """
         if permutation is None:
             return self._points_wocs_ctwcc
         return self._points_wocs_ctwcc.take(permutation, axis=0)
 
     def points_wcs_ctwocc(self, permutation=None):
+        """
+        :param permutation:
+        :return:
+        """
         if permutation is None:
             return self._points_wcs_ctwocc
         return np.concatenate((self._points_wcs_ctwocc[0:1], self._points_wocs_ctwocc.take(permutation, axis=0)))
 
     def points_wocs_ctwocc(self, permutation=None):
+        """
+        :param permutation:
+        :return:
+        """
         if permutation is None:
             return self._points_wocs_ctwocc
         return self._points_wocs_ctwocc.take(permutation, axis=0)
 
     @property
     def cn(self):
+        """
+        :return: Coordination number
+        """
         return len(self.coords)
 
     @property
     def coordination_number(self):
+        """
+        :return: Coordination number
+        """
         return len(self.coords)
 
 
@@ -414,8 +450,18 @@ class LocalGeometryFinder:
         self.setup_structure(
             Structure(lattice, species, coords, coords_are_cartesian))
 
-    def compute_coordination_environments(self, structure, indices=None, only_cations=True, strategy=DEFAULT_STRATEGY,
-                                          valences='bond-valence-analysis', initial_structure_environments=None):
+    def compute_coordination_environments(
+            self, structure, indices=None, only_cations=True, strategy=DEFAULT_STRATEGY,
+            valences='bond-valence-analysis', initial_structure_environments=None):
+        """
+        :param structure:
+        :param indices:
+        :param only_cations:
+        :param strategy:
+        :param valences:
+        :param initial_structure_environments:
+        :return:
+        """
         self.setup_structure(structure=structure)
         if valences == 'bond-valence-analysis':
             bva = BVAnalyzer()
@@ -673,9 +719,9 @@ class LocalGeometryFinder:
                                         to_add_from_hints.append({'isite': isite,
                                                                   'new_nb_set': new_nb_set,
                                                                   'cn_new_nb_set': cn_new_nb_set})
-                                        logging.debug('              => to be computed'.format(inew))
+                                        logging.debug('              => to be computed')
                                     else:
-                                        logging.debug('              => already present'.format(inew))
+                                        logging.debug('              => already present')
             logging.debug('    ... getting environments for nb_sets added from hints')
             for missing_nb_set_to_add in to_add_from_hints:
                 se.add_neighbors_set(isite=isite, nb_set=missing_nb_set_to_add['new_nb_set'])
@@ -711,6 +757,16 @@ class LocalGeometryFinder:
         return se
 
     def update_nb_set_environments(self, se, isite, cn, inb_set, nb_set, recompute=False, optimization=None):
+        """
+        :param se:
+        :param isite:
+        :param cn:
+        :param inb_set:
+        :param nb_set:
+        :param recompute:
+        :param optimization:
+        :return:
+        """
         ce = se.get_coordination_environments(isite=isite, cn=cn, nb_set=nb_set)
         if ce is not None and not recompute:
             return ce
@@ -790,6 +846,18 @@ class LocalGeometryFinder:
                                        random_rotation='NONE',
                                        random_scale='NONE',
                                        points=None):
+        """
+        :param symbol:
+        :param randomness:
+        :param max_random_dist:
+        :param symbol_type:
+        :param indices:
+        :param random_translation:
+        :param random_rotation:
+        :param random_scale:
+        :param points:
+        :return:
+        """
         if symbol_type == 'IUPAC':
             cg = self.allcg.get_geometry_from_IUPAC_symbol(symbol)
         elif symbol_type == 'MP' or symbol_type == 'mp_symbol':