+++ pymatgen/io/vasp/__init__.py
@@ -10,6 +10,6 @@ classes to be simply called as pymatgen.io.vasp.Incar for example, to retain
 backwards compatibility.
 """
 
-from .inputs import Incar, Poscar, Potcar, Kpoints, PotcarSingle
+from .inputs import Incar, Poscar, Potcar, Kpoints, PotcarSingle, VaspInput
 from .outputs import Vasprun, BSVasprun, Outcar, VolumetricData, Locpot, Chgcar, Elfcar, Procar, Oszicar, Xdatcar, \
     Dynmat, Wavecar, Wavederf, Waveder